Move model traits
