--- src/Models/Traits/WalletRequests.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Payavel\Checkout\Models\Traits;
-
-trait WalletRequests
-{
-    use ConfiguresCheckoutGateway;
-
-    /**
-     * Fetch the wallet details from the provider.
-     *
-     * @return \Payavel\Checkout\CheckoutResponse
-     */
-    public function fetch()
-    {
-        return $this->gateway->getWallet($this);
-    }
-}
